Added option to control whether or not artifacts are downloaded
There is a new option (downloadArtifacts) that controls whether or not
artifacts are downloaded. If they are (the default to maintain backwards
compatibility) then the trigger will check to see if any artifacts inside
a dependency have changed with out changing the version number.

If artifacts are not configured for download, the trigger only checks to
make sure the resolved dependency version has not changed.

Fixes JENKINS-28044