Fix FindBugs issue - Read & Write Ivy settings with UTF-8 encoding
It's better to use a defined encoding than to be platform-dependent.
Ivy settings are stored as XML which has UTF-8 as default encoding if
not otherwise specified in the file. Therefore using UTF-8.

An even better implementation would use an XML reader which parses the
content type information in the XML. But that's out of scope here.